🔥 Remove more unnecessary Pydantic v1 specific code
